+++ src/zimscraperlib/image/transformation.py
@@ -2,8 +2,9 @@
 # -*- coding: utf-8 -*-
 # vim: ai ts=4 sts=4 et sw=4 nu
 
+import io
 import pathlib
-from typing import Optional
+from typing import Optional, Union
 
 import PIL
 from resizeimage import resizeimage
@@ -13,17 +14,17 @@ from ..constants import ALPHA_NOT_SUPPORTED
 
 
 def resize_image(
-    src: pathlib.Path,
+    src: Union[pathlib.Path, io.BytesIO],
     width: int,
     height: Optional[int] = None,
-    dst: Optional[pathlib.Path] = None,
+    dst: Optional[Union[pathlib.Path, io.BytesIO]] = None,
     method: Optional[str] = "width",
     allow_upscaling: Optional[bool] = True,
     **params: Optional[dict],
 ) -> None:
-    """resize an image file to requested dimensions
+    """resize an image to requested dimensions
 
-    methods: width, height, cover
+    methods: width, height, cover, thumbnail
     allow upscaling: upscale image first, preserving aspect ratio if required"""
     with PIL.Image.open(src) as image:
         # preserve image format as resize() does not transmit it into new object
@@ -49,5 +50,9 @@ def resize_image(
     # remove alpha layer if not supported and added during resizing
     if resized.mode == "RGBA" and image_format in ALPHA_NOT_SUPPORTED:
         resized = resized.convert(image_mode)
+
+    # reset src if it's a byte stream and should be resized in-place
+    if dst is None and isinstance(src, io.BytesIO):
+        src.seek(0)
 
     save_image(resized, dst if dst is not None else src, image_format, **params)